sms: Mapper hints for games that are not correctly auto-detected
Some games attempt to write to addresses used by other mappers, causing
them to be misdirected. This patch adds mappers to the hints database.

+++ source/sms.h
@@ -59,7 +59,7 @@ typedef struct SMS_Context_s {
     uint32_t rom_size;
     uint32_t rom_mask;
     uint8_t  rom_hash [HASH_LENGTH];
-    uint8_t  rom_hints;
+    uint16_t rom_hints;
 
     uint8_t *bios;
     uint32_t bios_size;
